修正戰鬥相關函數的返回類型，改善代碼一致性；優化文本顯示邏輯，提升用戶體驗；調整視頻幀顯示邏輯，確保正確性

+++ src/fight.h
@@ -21,7 +21,7 @@
 #ifndef FIGHT_H
 #define FIGHT_H
 
-short PAL_BattleSelectAutoTargetFrom(
+int PAL_BattleSelectAutoTargetFrom(
     int);
 
 int PAL_IsPlayerDying(
